🚚 Rename: /home 디렉토리의 파일명 변경
