+++ src/components/home/AuctionOfTheMonth.vue
