+++ src/components/home/HotAuctions.vue
